fastapimainproject/requirements.txt
streamlit
pdfminer.six
python-docx
pandas
nltk